refactoring portfolio page
