+++ src/routes/app/portfolio/index.tsx
@@ -16,23 +16,11 @@ import {
 import { messagesContext } from "../layout";
 import {
   Form,
-  routeAction$,
-  routeLoader$,
-  server$,
-  z,
-  zod$,
 } from "@builder.io/qwik-city";
-import { connectToDB } from "~/utils/db";
-import jwt, { type JwtPayload } from "jsonwebtoken";
-import {
-  getDBTokenPriceUSD,
-  getResultAddresses,
-  getWalletDetails,
-} from "~/interface/wallets/observedWallets";
-import { type Wallet } from "~/interface/auth/Wallet";
+
 import { Modal } from "~/components/Modal/Modal";
 import { isValidName } from "~/utils/validators/addWallet";
-import { structureExists } from "~/interface/structure/removeStructure";
+
 import {
   simulateContract,
   writeContract,
@@ -44,8 +32,11 @@ import { getCookie } from "~/utils/refresh";
 import CoinsToTransfer from "~/components/Forms/portfolioTransfters/CoinsToTransfer";
 import CoinsAmounts from "~/components/Forms/portfolioTransfters/CoinsAmounts";
 import Destination from "~/components/Forms/portfolioTransfters/Destination";
-import { convertToFraction } from "../wallets";
 import { NoDataAdded } from "~/components/NoDataAdded/NoDataAdded";
+import { useDeleteStructure, useDeleteToken, useCreateStructure, useObservedWalletBalances, useAvailableStructures } from './server';
+export { useDeleteStructure, useDeleteToken, useCreateStructure, useObservedWalletBalances, useAvailableStructures } from './server';
+import { queryTokens } from "~/database/tokens";
+import { convertToFraction } from "~/utils/fractions";
 
 type WalletWithBalance = {
   wallet: {
@@ -74,250 +65,7 @@ export interface BatchTransferFormStore {
   receiverAddress: string;
   coinsToTransfer: StructureToApprove[];
 }
-export const useDeleteStructure = routeAction$(
-  async (structure, requestEvent) => {
-    const db = await connectToDB(requestEvent.env);
 
-    const cookie = requestEvent.cookie.get("accessToken");
-    if (!cookie) {
-      throw new Error("No cookie found");
-    }
-    if (!(await structureExists(db, structure.id))) {
-      throw new Error("Structure does not exist");
-    }
-
-    await db.delete(structure.id as string);
-
-    return {
-      success: true,
-      structure: { id: structure.id },
-    };
-  },
-  zod$({
-    id: z.string(),
-  }),
-);
-export const useDeleteToken = routeAction$(
-  async (data, requestEvent) => {
-    const db = await connectToDB(requestEvent.env);
-    const cookie = requestEvent.cookie.get("accessToken");
-    if (!cookie) {
-      throw new Error("No cookie found");
-    }
-    if (!(await structureExists(db, data.structureId))) {
-      throw new Error("Structure does not exist");
-    }
-
-    await db.query(`
-    DELETE structure_balance WHERE in=${data.structureId} AND out=${data.balanceId}`);
-
-    const [balanceCount]: any = await db.query(`
-    RETURN COUNT(SELECT id AS num_rows FROM structure_balance WHERE in=${data.structureId})`);
-
-    if (balanceCount === 0) {
-      await db.delete(data.structureId as string);
-    }
-  },
-  zod$({
-    structureId: z.string(),
-    balanceId: z.string(),
-  }),
-);
-
-export const useObservedWalletBalances = routeLoader$(async (requestEvent) => {
-  const db = await connectToDB(requestEvent.env);
-
-  const cookie = requestEvent.cookie.get("accessToken");
-  if (!cookie) {
-    throw new Error("No cookie found");
-  }
-  const { userId } = jwt.decode(cookie.value) as JwtPayload;
-
-  const resultAddresses = await getResultAddresses(db, userId);
-  if (!resultAddresses[0]) {
-    return [];
-  }
-  const walletsWithBalance = [];
-  let balanceDetails = [];
-
-  for (const observedWalletAddress of resultAddresses) {
-    const walletDetails = await getWalletDetails(
-      db,
-      observedWalletAddress.address,
-      userId,
-    );
-    const [balances]: any = await db.query(
-      `SELECT id, value FROM balance WHERE ->(for_wallet WHERE out = '${walletDetails.id}')`,
-    );
-
-    const walletNameResult: any = await db.query(
-      `SELECT VALUE name FROM ${walletDetails.id}<-observes_wallet WHERE in = ${userId}`,
-    );
-
-    for (const balance of balances) {
-      if (balance.value === "0") {
-        continue;
-      }
-      const [tokenId]: any = await db.query(`
-      SELECT VALUE ->for_token.out FROM ${balance.id}`);
-
-      const [tokenDetails]: any = await db.query(`
-      SELECT * FROM ${tokenId[0]}`);
-
-      const walletBalance = {
-        balanceId: balance.id,
-        tokenId: tokenDetails[0].id,
-        tokenSymbol: tokenDetails[0].symbol,
-      };
-      balanceDetails.push(walletBalance);
-    }
-
-    const walletWithBalance = {
-      wallet: walletDetails,
-      walletName: walletNameResult,
-      balance: balanceDetails,
-    };
-    balanceDetails = [];
-
-    walletsWithBalance.push(walletWithBalance);
-  }
-  return walletsWithBalance;
-});
-
-export const useAvailableStructures = routeLoader$(async (requestEvent) => {
-  const db = await connectToDB(requestEvent.env);
-  const cookie = requestEvent.cookie.get("accessToken");
-  if (!cookie) {
-    throw new Error("No cookie found");
-  }
-  const { userId } = jwt.decode(cookie.value) as JwtPayload;
-
-  const [result]: any = await db.query(`
-    SELECT VALUE ->has_structure.out FROM ${userId}`);
-
-  if (!result) throw new Error("No structures available");
-  const createdStructureQueryResult = result[0];
-  const availableStructures: any[] = [];
-
-  for (const createdStructure of createdStructureQueryResult) {
-    const [structure] = await db.select(`${createdStructure}`);
-    const structureTokens: any = [];
-    const [structureBalances]: any = await db.query(`
-      SELECT VALUE ->structure_balance.out
-      FROM ${structure.id}`);
-
-    if (!structureBalances[0].length) {
-      await db.delete(structure.id);
-    } else {
-      for (const balance of structureBalances[0]) {
-        const [walletId]: any = await db.query(`
-        SELECT out
-        FROM for_wallet
-        WHERE in = ${balance}`);
-
-        const [wallet] = await db.select<Wallet>(`${walletId[0].out}`);
-
-        const [[walletNameResult]]: any = await db.query(
-          `SELECT VALUE name FROM ${wallet.id}<-observes_wallet WHERE in = ${userId}`,
-        );
-
-        const [tokenBalance]: string[] = await db.query(`
-        SELECT VALUE value
-        FROM balance
-        WHERE id = ${balance}`);
-
-        const [tokenId]: any = await db.query(`
-        SELECT VALUE ->for_token.out
-        FROM ${balance}`);
-
-        const [token]: any = await db.query(
-          `SELECT *
-         FROM ${tokenId[0]}`,
-        );
-
-        const [tokenValue] = await getDBTokenPriceUSD(db, token[0].address);
-        const tokenWithBalance = {
-          id: token[0].id,
-          name: token[0].name,
-          symbol: token[0].symbol,
-          decimals: token[0].decimals,
-          balance: tokenBalance[0],
-          balanceValueUSD: tokenValue.priceUSD,
-          balanceId: balance,
-        };
-
-        structureTokens.push({
-          wallet: {
-            id: wallet.id,
-            name: walletNameResult,
-            chainId: wallet.chainId,
-            isExecutable: wallet.isExecutable,
-          },
-          balance: tokenWithBalance,
-        });
-      }
-
-      availableStructures.push({
-        structure: {
-          id: structure.id,
-          name: structure.name,
-        },
-        structureBalance: structureTokens,
-      });
-    }
-  }
-
-  return availableStructures;
-});
-export const useCreateStructure = routeAction$(
-  async (data, requestEvent) => {
-    const cookie = requestEvent.cookie.get("accessToken");
-    if (!cookie) {
-      throw new Error("No cookie found");
-    }
-    const { userId } = jwt.decode(cookie.value) as JwtPayload;
-
-    const db = await connectToDB(requestEvent.env);
-
-    const [namesList]: any = await db.query(`
-    SELECT VALUE out.name FROM ${userId}->has_structure`);
-
-    if (namesList.includes(data.name)) {
-      return {
-        success: false,
-        message: "Name already taken",
-      };
-    }
-    const structure = await db.create("structure", {
-      name: data.name,
-    });
-
-    await db.query(`
-    relate only ${userId}-> has_structure -> ${structure[0].id}`);
-
-    for (const balanceId of data.balancesId) {
-      await db.query(`
-      relate only ${structure[0].id}-> structure_balance -> ${balanceId}`);
-    }
-
-    return {
-      success: true,
-      structure: { name: data.name, balances: data.balancesId },
-    };
-  },
-  zod$({
-    name: z.string().min(2, { message: "structure name invalid" }),
-    balancesId: z.array(z.string()),
-  }),
-);
-const queryTokens = server$(async function () {
-  const db = await connectToDB(this.env);
-
-  const [tokens]: any = await db.query(
-    `SELECT decimals, symbol, address FROM token;`,
-  );
-  return tokens;
-});
 export default component$(() => {
   const modalStore = useContext(ModalStoreContext);
   const clickedToken = useStore({ balanceId: "", structureId: "" });
