+++ src/routes/app/wallets/index.tsx
@@ -11,7 +11,7 @@ import {
 import { Form, server$ } from "@builder.io/qwik-city";
 import { type JwtPayload } from "jsonwebtoken";
 import { contractABI } from "~/abi/abi";
-import { connectToDB } from "~/utils/db";
+import { connectToDB } from "~/database/db";
 import { chainIdToNetworkName } from "~/utils/chains";
 import { Modal } from "~/components/Modal/Modal";
 import { SelectedWalletDetails } from "~/components/Wallets/Details/SelectedWalletDetails";
@@ -53,41 +53,13 @@ import { EvmChain } from "@moralisweb3/common-evm-utils";
 import { useAddWallet } from "./server/addWalletAction";
 import { useRemoveWallet } from "./server/removeWalletAction";
 import { useGetBalanceHistory } from "./server/getBalanceHistoryAction";
+import { chekckIfProperAmount, convertToFraction, replaceNonMatching } from "~/utils/fractions";
 export { useAddWallet } from "./server/addWalletAction";
 export { useRemoveWallet } from "./server/removeWalletAction";
 export { useGetBalanceHistory } from "./server/getBalanceHistoryAction";
 
-export const convertToFraction = (numericString: string) => {
-  let fractionObject;
-  if (!numericString.includes(".")) {
-    fractionObject = {
-      numerator: BigInt(numericString),
-      denominator: BigInt(1),
-    };
-  } else {
-    const fractionArray = numericString.split(".");
-    fractionObject = {
-      numerator: BigInt(`${fractionArray[0]}${fractionArray[1]}`),
-      denominator: BigInt(Math.pow(10, fractionArray[1].length)),
-    };
-  }
-  return fractionObject;
-};
 
-export function replaceNonMatching(
-  inputString: string,
-  regex: RegExp,
-  replacement: string,
-): string {
-  return inputString.replace(
-    new RegExp(`[^${regex.source}]`, "g"),
-    replacement,
-  );
-}
 
-export const chekckIfProperAmount = (input: string, regex: RegExp) => {
-  return regex.test(input);
-};
 
 export interface addWalletFormStore {
   name: string;
