+++ src/routes/app/wallets/server/getBalanceHistoryAction.ts
@@ -1,6 +1,6 @@
 import { routeAction$ } from "@builder.io/qwik-city";
 import { getErc20TokenTransfers, getWalletBalance } from "~/server/moralis";
-import { connectToDB } from "~/utils/db";
+import { connectToDB } from "~/database/db";
 import { convertWeiToQuantity } from "~/utils/formatBalances/formatTokenBalance";
 
 /**
