+++ src/utils/validators/addWallet.ts
@@ -1,5 +1,5 @@
 import { isAddress, getAddress } from "viem";
-import { connectToDB } from "../db";
+import { connectToDB } from "../../database/db";
 import { server$, z } from "@builder.io/qwik-city";
 import jwt, { type JwtPayload } from "jsonwebtoken";
